latent-assets/backend/pyproject.toml
[project]
name = "latent-assets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.8",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
]
